Fixing circular import
