+++ dgo/tests.py
@@ -1,10 +1,10 @@
 from django.test import TestCase
-from .models import GOD, GODPort, GODPortConnectionType, Jumper
+from .models import GOD, GODPort, GODPortConnectionType, Jumper, AccessCable
 from rest_framework.test import APIRequestFactory
 from gbic.models import GBIC, GBIC_Type
 from ipa.models import Site, SiteType, ParticipantInstitution, InstitutionType
 from .views import GODListViewSet, GODPortConnectionTypeListViewSet, GODPortListViewSet
-from .views import JumperViewSet
+from .views import JumperViewSet, AccessCableListViewSet
 
 class GODTest(TestCase):
 
@@ -114,7 +114,18 @@ class JumperTest(TestCase):
         tipo = GODPortConnectionType.objects.create(code = "abcd")
         gbictipo = GBIC_Type.objects.create(description = "tipoteste")
         gbic = GBIC.objects.create(serial = "huehue", patrimony_number = "123", gbic_type = gbictipo)
-        dgo = GOD.objects.create(code = 198, fabricant="HotAntardida", port_quantity=42)
+        ipatype = InstitutionType.objects.create(description='BestEver')
+        ipa = ParticipantInstitution.objects.create(name='Uns Paranauês Maneiros', institution_type=ipatype)
+        sitetype = SiteType.objects.create(description='IDK What Am I Doing')
+        sitetest = Site.objects.create(
+            name='Lenny Face',
+            lattitude=42.42,
+            longitude=54.54,
+            bandwidth=42,
+            ipa_code=ipa,
+            site_type=sitetype
+        )
+        dgo = GOD.objects.create(code = 198, fabricant="HotAntardida", port_quantity=42, site_id=sitetest)
         GODPorttest1 = GODPort.objects.create(code=999, connection_type=tipo, god_id=dgo, gbic_id=gbic)
         GODPorttest2 = GODPort.objects.create(code=109, connection_type=tipo, god_id=dgo, gbic_id=gbic)
         jumper_test = Jumper.objects.create(god_port1=GODPorttest1, god_port2=GODPorttest2)
@@ -127,7 +138,18 @@ class JumperTest(TestCase):
         tipo = GODPortConnectionType.objects.create(code = "abcd")
         gbictipo = GBIC_Type.objects.create(description = "tipoteste")
         gbic = GBIC.objects.create(serial = "huehue", patrimony_number = "123", gbic_type = gbictipo)
-        dgo = GOD.objects.create(code = 198, fabricant="HotAntardida", port_quantity=42)
+        ipatype = InstitutionType.objects.create(description='BestEver')
+        ipa = ParticipantInstitution.objects.create(name='Uns Paranauês Maneiros', institution_type=ipatype)
+        sitetype = SiteType.objects.create(description='IDK What Am I Doing')
+        sitetest = Site.objects.create(
+            name='Lenny Face',
+            lattitude=42.42,
+            longitude=54.54,
+            bandwidth=42,
+            ipa_code=ipa,
+            site_type=sitetype
+        )
+        dgo = GOD.objects.create(code = 198, fabricant="HotAntardida", port_quantity=42, site_id=sitetest)
         GODPorttest1 = GODPort.objects.create(code=999, connection_type=tipo, god_id=dgo, gbic_id=gbic)
         GODPorttest2 = GODPort.objects.create(code=109, connection_type=tipo, god_id=dgo, gbic_id=gbic)
         jumper_test = Jumper.objects.create(god_port1=GODPorttest1, god_port2=GODPorttest2)
@@ -135,3 +157,33 @@ class JumperTest(TestCase):
         jumper_test.delete()
         response = jumper_detail(request, pk=jumper_test.pk)
         self.assertEqual(response.status_code, 404)
+
+
+class AccessCableTest(TestCase):
+
+    def test_acess_cable_view_set(self):
+        request = APIRequestFactory().get("")
+        view = AccessCableListViewSet.as_view(actions={'get': 'list'})
+        site_type = SiteType.objects.create(description="RandomSiteType")
+        instituion_type = InstitutionType.objects.create(description="RandomInstitution")
+        ipa = ParticipantInstitution.objects.create(name='UnB', institution_type=instituion_type)
+        site = Site.objects.create(name='RandomSite', lattitude=42, longitude=42, bandwidth=42, ipa_code=ipa, site_type=site_type)
+        god = GOD.objects.create(code=1, fabricant="FabricanteTeste", port_quantity="10", site_id=site)
+        access_cable = AccessCable.objects.create(cod=1, length=120, fiber_quantity=10, god_id=god, site_id=site)
+
+        response = view(request, pk=access_cable.pk)
+        self.assertEqual(response.status_code, 200)
+
+    def test_access_cable_post(self):
+        view = AccessCableListViewSet.as_view(actions={'get': 'list'})
+        factory = APIRequestFactory()
+        site_type = SiteType.objects.create(description="RandomSiteType")
+        instituion_type = InstitutionType.objects.create(description="RandomInstitution")
+        ipa = ParticipantInstitution.objects.create(name='UnB', institution_type=instituion_type)
+        site = Site.objects.create(name='RandomSite', lattitude=42, longitude=42, bandwidth=42, ipa_code=ipa, site_type=site_type)
+        god = GOD.objects.create(code=1, fabricant="FabricanteTeste", port_quantity="10",site_id=site)
+        access_cable = AccessCable.objects.create(cod=1, length=120, fiber_quantity=10, god_id=god, site_id=site)
+        request = factory.get('/access-cables/')
+        response = view(request)
+        items = response.data[0]
+        self.assertEqual(list(items.values())[1], 120)
