+++ ipa/serializers.py
@@ -1,9 +1,8 @@
 from .models import InstitutionType, ParticipantInstitution
-from .models import ContactType, Generator, AccessCable
+from .models import ContactType, Generator
 from .models import NoBreak
 from .models import Switch
 from .models import Contact
-from dgo.serializers import GODSerializer
 from sigi_op.serializers import SiteSerializer
 from rest_framework import serializers
 
@@ -89,19 +88,4 @@ class SwitchSerializer(serializers.ModelSerializer):
                 'slots_quantity',
                 'patrimony_number',
                 'site_id',
-        ]
-
-
-class AccessCableSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AccessCable
-        god_id = GODSerializer(many=True, read_only=True)
-        site_id = SiteSerializer(many=True, read_only=True)
-        fields = [
-            'id',
-            'length',
-            'fiber_quantity',
-            'cod',
-            'site_id',
-            'god_id',
         ]
